Remove TODO note related to 'countable' property that could potentially be returned to API clients
This comment was fairly tricky to reason about after re-reading months/years later, but basically it's asking whether the client should be provided with a flag to determine whether a given ingredient name is 'countable' or not.  In other words: should the client check how many of the ingredient are mentioned in the given context ('1 tomato', '2 tomatoes'), and if it _is_ a countable ingredient, then use the plural form (or not, as appropriate).  On reflection I'm not sure that makes much sense: if it's not a countable ingredient, then product.singular and product.plural should contain the same names anyway -- and frankly in the wider scheme of future internationalization, countable vs not-countable is a surface detail compared to the number of possible name-rendering-context problems